move filter to module
