+++ src/main/kotlin/net/ccbluex/liquidbounce/utils/block/bed/BedBlockTracker.kt
@@ -67,30 +67,6 @@ object BedBlockTracker : AbstractBlockLocationTracker.BlockPos2State<BedState>()
         triggerRescan()
     }
 
-    private val WHITELIST_NON_SOLID: Set<Block> = ReferenceOpenHashSet.of(
-        Blocks.LADDER,
-
-        Blocks.WATER,
-
-        Blocks.GLASS,
-        Blocks.WHITE_STAINED_GLASS,
-        Blocks.ORANGE_STAINED_GLASS,
-        Blocks.MAGENTA_STAINED_GLASS,
-        Blocks.LIGHT_BLUE_STAINED_GLASS,
-        Blocks.YELLOW_STAINED_GLASS,
-        Blocks.LIME_STAINED_GLASS,
-        Blocks.PINK_STAINED_GLASS,
-        Blocks.GRAY_STAINED_GLASS,
-        Blocks.LIGHT_GRAY_STAINED_GLASS,
-        Blocks.CYAN_STAINED_GLASS,
-        Blocks.PURPLE_STAINED_GLASS,
-        Blocks.BLUE_STAINED_GLASS,
-        Blocks.BROWN_STAINED_GLASS,
-        Blocks.GREEN_STAINED_GLASS,
-        Blocks.RED_STAINED_GLASS,
-        Blocks.BLACK_STAINED_GLASS,
-    )
-
     private fun BlockPos.getBedSurroundingBlocks(blockState: BlockState): List<SurroundingBlock> {
         val layers = Array<Reference2IntOpenHashMap<Block>>(maxLayers) { Reference2IntOpenHashMap() }
 
@@ -103,11 +79,8 @@ object BedBlockTracker : AbstractBlockLocationTracker.BlockPos2State<BedState>()
                 continue
             }
 
-            val block = state.block
-            if (state.isSolidBlock(world, pos) || block in WHITELIST_NON_SOLID) {
-                // Count blocks (default = 0)
-                layers[layer - 1].addTo(block, 1)
-            }
+            // Count blocks (default = 0)
+            layers[layer - 1].addTo(state.block, 1)
         }
 
         val result = arrayOfNulls<SurroundingBlock>(layers.sumOf { it.size })
